apps/api/pyproject.toml
[project]
name = "serverless-core-api"
version = "0.1.0"
description = "serverless-core FastAPI control plane"
requires-python = ">=3.12,<3.14"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn[standard]>=0.34",
  "httpx>=0.28",
  "pydantic>=2.9,<3.0",
  "pydantic-settings>=2.6",
  "supabase>=2.10",
  "pyjwt[crypto]>=2.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/serverless_core_api"]

[dependency-groups]
dev = [
  "pytest>=8.3",
  "pytest-asyncio>=0.24",
  "respx>=0.21",
  "ruff>=0.7",
]
